Change type statistics variables

+++ src/main/java/controllers/StatisticsController.java
@@ -46,9 +46,9 @@ public class StatisticsController implements Initializable {
 
         }
         else {
-            int overallVelocityVal = statisticsModel.getOverallVelocity();
-            int leadTimeVal = statisticsModel.getLeadTime();
-            int averageWIPVal = statisticsModel.getAverageWIP(statisticsModel.getActiveWeek());
+            double overallVelocityVal = statisticsModel.getOverallVelocity();
+            double leadTimeVal = statisticsModel.getLeadTime();
+            double averageWIPVal = statisticsModel.getAverageWIP(statisticsModel.getActiveWeek());
 
             if(overallVelocityVal == -1) overallVelocity.setText("There's no story points on the Completed Work column yet");
             else overallVelocity.setText(overallVelocityVal + " story points per week");
@@ -68,7 +68,7 @@ public class StatisticsController implements Initializable {
         XYChart.Series ltSeries = new XYChart.Series();
         XYChart.Series wipSeries = new XYChart.Series();
 
-        int week = statisticsModel.getActiveWeek();
+        double week = statisticsModel.getActiveWeek();
         System.out.println("week: ");
         System.out.println(week);
